requirements.txt
pymongo==4.6.3
discord==2.3.2
keep_alive==0.0.1
python-dotenv
motor==3.4.0